Refactor: 리팩토링
- 사용하지 않는 코드 삭제
- 중복 코드 제거
- 권한이 없는 사용자는 수정, 삭제 불가능하도록 변경
- 명칭 공통화

+++ src/main/java/com/pinterest/service/FollowService.java
@@ -54,10 +54,14 @@ public class FollowService {
     }
 
     @Transactional
-    public void cancel(Long followId) {
+    public void cancel(Long followId, String loginEmail) {
         Follow follow = followRepository.findById(followId)
                 .orElseThrow(() -> new EntityNotFoundException("팔로우 정보가 없습니다."));
-        followRepository.delete(follow);
+        if (follow.getFromMember().getEmail().equals(loginEmail)) {
+            followRepository.delete(follow);
+        } else {
+            throw new EntityNotFoundException("권한이 없습니다.");
+        }
     }
 
     public List<MemberDto> getFollowers(String email) {
